Parse terraform state instead of main.tf.json when possible
This fixes the UX bug that the cluster editor would show the planned configuration,
instead of the currently applied configuration, which was misleading. However, because
parsing the terraform state file is much slower than parsing the main.tf.json file,
we only use the terraform state file when accessing a single cluster's configuration,
not when the user is listing all clusters (on the home page for instance).

This reverts a good part of the commit https://github.com/ComputeCanada/mc-hub/commit/ad6e4779889751e312b6b3d8274da8303a840667.